requirements.txt
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
idna==3.3
playsound==1.3.0
PyBluez==0.23
pycairo==1.20.1
pycparser==2.21
pydub==0.25.1
PyGObject==3.42.0
PyYAML==6.0
requests==2.27.1
requests-futures==1.0.0
scapy==2.4.5
sounddevice==0.4.4
urllib3==1.26.8
